Moves label_extraction
