+++ kglib/kgcn/encode/boolean.py
