+++ kglib/kgcn/encode/encode_test.py
@@ -23,7 +23,7 @@ import grakn
 import numpy as np
 import tensorflow as tf
 
-import kglib.kgcn.encoder.encode as encode
+import kglib.kgcn.encode.encode as encode
 
 
 class TestEncode(unittest.TestCase):
