+++ kglib/kgcn/encode/schema.py
