+++ kglib/kgcn/management/samples.py
@@ -21,8 +21,8 @@ import collections
 
 import numpy as np
 
-from kglib.kgcn.neighbourhood.data.sampling import random_sampling as random
-from kglib.kgcn.use_cases.attribute_prediction import label_extraction as label_extraction
+import kglib.kgcn.neighbourhood.data.sampling.random_sampling as random
+import kglib.kgcn.learn.classify.attribute.label_extraction as label_extraction
 
 
 def query_for_random_samples_with_attribute(tx, query, example_var_name, attribute_var_name, attribute_vals,
